tests added

+++ project/src/main/java/ru/price_in_words/api/services/Parser.java
@@ -3,5 +3,5 @@ package ru.price_in_words.api.services;
 import ru.price_in_words.domain.Numbers;
 
 public interface Parser {
-    Numbers parseInt(String stringNumber);
+    Numbers parseLong(String stringNumber);
 }
